+++ NFIQ2/NFIQ2Algorithm/include/nfiq2/timer.hpp
