+++ NFIQ2/NFIQ2Algorithm/src/features/FDAFeature.cpp
@@ -5,7 +5,7 @@
 #include "FeatureFunctions.h"
 #include "include/nfiq2/nfiqexception.h"
 
-#include "include/Timer.hpp"
+#include "nfiq2/timer.hpp"
 #include <sstream>
 
 #if defined WINDOWS || defined WIN32
