+++ NFIQ2/NFIQ2Algorithm/src/features/FingerJetFXFeature.cpp
@@ -3,7 +3,7 @@
 #include "FingerJetFXFeature.h"
 #include "include/nfiq2/nfiqexception.h"
 
-#include "include/Timer.hpp"
+#include "nfiq2/timer.hpp"
 #include <sstream>
 
 #define CBEFF (0x00330502)
