+++ NFIQ2/NFIQ2Algorithm/src/features/ImgProcROIFeature.cpp
@@ -3,7 +3,7 @@
 #include "ImgProcROIFeature.h"
 #include "include/nfiq2/nfiqexception.h"
 
-#include "include/Timer.hpp"
+#include "nfiq2/timer.hpp"
 #include <sstream>
 
 using namespace NFIQ;
