+++ NFIQ2/NFIQ2Algorithm/src/features/LCSFeature.cpp
@@ -4,7 +4,7 @@
 #include "LCSFeature.h"
 #include "include/nfiq2/nfiqexception.h"
 
-#include "include/Timer.hpp"
+#include "nfiq2/timer.hpp"
 #include <iostream>
 #include <sstream>
