+++ NFIQ2/NFIQ2Algorithm/src/features/MuFeature.cpp
@@ -1,7 +1,7 @@
 #include "MuFeature.h"
 #include "include/nfiq2/nfiqexception.h"
 
-#include "include/Timer.hpp"
+#include "nfiq2/timer.hpp"
 #include <sstream>
 
 using namespace NFIQ;
