+++ NFIQ2/NFIQ2Algorithm/src/features/OCLHistogramFeature.cpp
@@ -2,7 +2,7 @@
 #include "OCLHistogramFeature.h"
 #include "include/nfiq2/nfiqexception.h"
 
-#include "include/Timer.hpp"
+#include "nfiq2/timer.hpp"
 #include <sstream>
 
 #if defined WINDOWS || defined WIN32
