+++ NFIQ2/NFIQ2Algorithm/src/features/OFFeature.cpp
@@ -2,7 +2,7 @@
 #include "OFFeature.h"
 #include "include/nfiq2/nfiqexception.h"
 
-#include "include/Timer.hpp"
+#include "nfiq2/timer.hpp"
 #include <sstream>
 //#include <iostream>
 #include <math.h>
