+++ NFIQ2/NFIQ2Algorithm/src/features/RVUPHistogramFeature.cpp
@@ -4,7 +4,7 @@
 #include "RVUPHistogramFeature.h"
 #include "include/nfiq2/nfiqexception.h"
 
-#include "include/Timer.hpp"
+#include "nfiq2/timer.hpp"
 #include <cmath>
 #include <sstream>
