+++ NFIQ2/NFIQ2Algorithm/src/nfiq2/nfiq2impl.cpp
@@ -14,7 +14,7 @@
 #include "include/nfiq2/nfiqexception.h"
 #include "nfiq2impl.h"
 
-#include "include/Timer.hpp"
+#include "nfiq2/timer.hpp"
 #include <iomanip>
 #include <iostream>
 #include <list>
